+++ toys/lib/toys/templates/rdoc.rb
@@ -30,61 +30,92 @@
 module Toys
   module Templates
     ##
-    # A template that generates yardoc tools.
+    # A template that generates rdoc tools.
     #
-    class Yardoc
+    class Rdoc
       include Template
 
       ##
-      # Default version requirements for the yard gem.
-      # @return [String]
+      # Default version requirements for the rdoc gem.
+      # @return [Array<String>]
       #
-      DEFAULT_GEM_VERSION_REQUIREMENTS = "~> 0.9".freeze
+      DEFAULT_GEM_VERSION_REQUIREMENTS = ">= 5.0.0".freeze
 
       ##
       # Default tool name
       # @return [String]
       #
-      DEFAULT_TOOL_NAME = "yardoc".freeze
+      DEFAULT_TOOL_NAME = "rdoc".freeze
 
       ##
-      # Create the template settings for the Yardoc template.
+      # Default output directory
+      # @return [String]
+      #
+      DEFAULT_OUTPUT_DIR = "html".freeze
+
+      ##
+      # Create the template settings for the Rdoc template.
       #
       # @param [String] name Name of the tool to create. Defaults to
       #     {DEFAULT_TOOL_NAME}.
       # @param [String,Array<String>] gem_version Version requirements for
-      #     the yard gem. Defaults to {DEFAULT_GEM_VERSION_REQUIREMENTS}.
+      #     the rdoc gem. Defaults to {DEFAULT_GEM_VERSION_REQUIREMENTS}.
       # @param [Array<String>] files An array of globs indicating the files
       #     to document.
-      # @param [Array<String>] options Additional options passed to YARD
-      # @param [Array<String>] stats_options Additional options passed to YARD
-      #     stats
+      # @param [String] output_dir Name of directory to receive html output
+      #     files. Defaults to {DEFAULT_OUTPUT_DIR}.
+      # @param [String,nil] markup Markup format. Allowed values include
+      #     "rdoc", "rd", and "tomdoc". Default is "rdoc".
+      # @param [String,nil] title Title of RDoc documentation. If `nil`, RDoc
+      #     will use a default title.
+      # @param [String,nil] main Name of the file to use as the main top level
+      #     document. Default is none.
+      # @param [String,nil] template Name of the template to use. If `nil`,
+      #     RDoc will use its default template.
+      # @param [String,nil] generator Name of the format generator. If `nil`,
+      #     RDoc will use its default generator.
+      # @param [Array<String>] options Additional options to pass to RDoc.
       #
       def initialize(name: nil,
                      gem_version: nil,
                      files: [],
-                     options: [],
-                     stats_options: [])
+                     output_dir: nil,
+                     markup: nil,
+                     title: nil,
+                     main: nil,
+                     template: nil,
+                     generator: nil,
+                     options: [])
         @name = name || DEFAULT_TOOL_NAME
         @gem_version = gem_version || DEFAULT_GEM_VERSION_REQUIREMENTS
         @files = files
+        @output_dir = output_dir || DEFAULT_OUTPUT_DIR
+        @markup = markup
+        @title = title
+        @main = main
+        @template = template
+        @generator = generator
         @options = options
-        @stats_options = stats_options
       end
 
       attr_accessor :name
       attr_accessor :gem_version
       attr_accessor :files
+      attr_accessor :output_dir
+      attr_accessor :markup
+      attr_accessor :title
+      attr_accessor :main
+      attr_accessor :template
+      attr_accessor :generator
       attr_accessor :options
-      attr_accessor :stats_options
 
       to_expand do |template|
         tool(template.name) do
-          desc "Run yardoc on the current project."
+          desc "Run rdoc on the current project."
 
           run do
-            ::Toys::Utils::Gems.activate("yard", *Array(template.gem_version))
-            require "yard"
+            ::Toys::Utils::Gems.activate("rdoc", *Array(template.gem_version))
+            require "rdoc"
 
             files = []
             patterns = Array(template.files)
@@ -94,16 +125,15 @@ module Toys
             end
             files.uniq!
 
-            unless template.stats_options.empty?
-              template.options << "--no-stats"
-              template.stats_options << "--use-cache"
-            end
+            args = template.options.dup
+            args << "-o" << template.output_dir
+            args << "--markup" << template.markup if template.markup
+            args << "--main" << template.main if template.main
+            args << "--title" << template.title if template.title
+            args << "-T" << template.template if template.template
+            args << "-f" << template.generator if template.generator
 
-            yardoc = ::YARD::CLI::Yardoc.new
-            yardoc.run(*(template.options + files))
-            unless template.stats_options.empty?
-              ::YARD::CLI::Stats.run(*template.stats_options)
-            end
+            RDoc::RDoc.new.document(args + files)
           end
         end
       end